Implement lic_0
This adds an implementation for the lic_0 function.
It returns True when two consecutive points have a distance
larger than the parameter "length1" and false otherwise.

Two unit tests have also been added.
The first test checks the case when the condition is true
and the second test when the condition is false.