Implement lic_2
This adds an implementation for the lic_2 function.
It returns true when the angle satisfies angle > (PI-EPSILON)
(or equivalently when the outer angle satisfies angle < (PI+EPSILON)).

This also adds a set of unit tests.
There is a testcase for when the condition is true and another for when
it is false.

There are also 4 test cases that handle the edge case when the first or
the last point in a sequence coincides
with the vertex.