Implement lic_1
This adds an implementation for the lic_1 function. It returns true
when any three consecutive data points cannot all be contained inside
a circle with a radius of RADIUS1. This also adds 4 unit tests that
cover the case when the statement is true and false both for points
that form non-degenerate and degenerate triangles.

Fixes #5