Implement lic_12
The lic_12 function checks if there are enough points (at least 3) and if the "length2" parameter is positive.
If either of these conditions is not met, it returns False. Otherwise, it iterates over pairs of points in the "points" list, checking the distance between each pair
If there's a pair of points that have a distance greater than "length1" and less than "length2", it returns True, otherwise it returns False.

fixes #18